requirements.txt
flask==3.1.3
flask-login==0.6.3
werkzeug==3.1.8
ortools==9.15.6755
requests==2.33.1
python-dotenv==1.2.2
gunicorn==25.3.0
psycopg2-binary==2.9.11
anthropic==0.93.0
